+++ tests/assets/og_images/_data.js
@@ -1,0 +1,1 @@
+export { default as ogImageTemplate } from "./_includes/og_template.jsx";